requirements.txt
asn1crypto==0.24.0
certifi==2018.10.15
cffi==1.11.5
chardet==3.0.4
Click==7.0
cryptography==2.4.2
Deprecated==1.2.4
Flask==1.0.2
idna==2.7
itsdangerous==1.1.0
Jinja2==2.10
MarkupSafe==1.1.0
pycparser==2.19
PyGithub==1.43.3
PyJWT==1.6.4
requests==2.20.1
six==1.11.0
urllib3==1.24.1
Werkzeug==0.14.1
wrapt==1.10.11
